function-source/requirements.txt
# Function dependencies, for example:
# package>=version
pandas>=1.2.0
bs4>=0.0.1
google-cloud-bigquery>=1.24.0
numpy>=1.19.1
google-auth>=1.23.0
tqdm>=4.56.0
pandas-gbq>=0.15.0